security: M20 — overflow-safe directory offset computation in GRP/HOG/GOB/POD engines
GrpEngine::Open computed the data-region start offset as
16 + count*16 entirely in uint32_t, where count is read directly from the
file header; a crafted GRP claiming a huge entry count wraps this to a
small value, corrupting every subsequent entry offset. Now computed in
uint64_t and rejected if it doesn't fit back in uint32_t.

HogEngine/GobEngine/PodEngine read per-entry offsets directly from their
directory records (no attacker-controlled count-based offset computation on
the read path), but doSave in all three computed the write-side starting
data offset the same overflowing way from m_entries.size(); switched to the
same uint64_t-then-cast pattern for consistency and defense in depth.

+++ src/engine/GobEngine.cpp
@@ -47,7 +47,7 @@ bool GobEngine::doSave(std::ofstream& f)
     f.write("GOB ", 4);
     writeLE32(f, static_cast<uint32_t>(count));
 
-    uint32_t dataOff = 8 + static_cast<uint32_t>(count) * 136;
+    uint32_t dataOff = static_cast<uint32_t>(8ull + static_cast<uint64_t>(count) * 136ull);
 
     // Write directory entries (offset[4] + size[4] + name[128])
     for (const auto& e : m_entries)
